flogging

Macro const_logger

macro_rules! const_logger {
    ($block:block) => { ... };
}

Setup module level logger access.

The basic macro syntax is:

const_logger!({/* the block of Rust code to build a Logger goes here */});

Notice there are curly braces “{}” wrapping the inner Rust code. They are required.

The code you put in here will depend on what configuration of Logger you want to setup.

Examples

extern crate flogging;
use flogging::*;

const_logger!({
    Logger::builder(module_path!())
        .set_level(Level::FINEST)
        .add_console_handler()
        .add_file_handler_with("rdb.log", FormatType::Iso8601, None)
        .build()
});